AMP-373: fixing all tests
