+++ src/main/java/com/github/jmchilton/blend4j/galaxy/beans/WorkflowOutputs.java
@@ -4,22 +4,24 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+import org.codehaus.jackson.annotate.JsonProperty;
 
 /**
  * AMPPD extension
  * Bean for fields returned by Galaxy upon workflow invocation, which contains inputs/outputs and basic information of each invocation step without details about jobs/outputs. 
  */
 @JsonIgnoreProperties(ignoreUnknown=true)
-public class WorkflowOutputs extends InvocationBase {
+public class WorkflowOutputs extends InvocationBriefs {
 	
-	private List<InvocationStep> steps = new ArrayList<InvocationStep>();
+	private List<String> outputIds = new ArrayList<String>(); // correspond to "output" field
 
-	public List<InvocationStep> getSteps() {
-		return steps;
+	public List<String> getOutputIds() {
+		return outputIds;
 	}
 
-	public void setSteps(List<InvocationStep> steps) {
-		this.steps = steps;
-	}
+	@JsonProperty("outputs")
+	public void setOutputIds(final List<String> outputIds) {
+		this.outputIds = outputIds;
+	}	
 	
 }
